put some tests in test_sort.py into another two files - test_push_based_shuffle.py and test_shuffle_diagnostics.py to enhance CI parallelization
